LPS-193648 refactor TempAssetCategoryAttachmentsUploadFileEntryHandler.java and AssetCategoryAttachmentsUploadResponseHandler.java
move and create inner class in UploadTempAssetCategoryAttachmentMVCActionCommand.java and delete handler class
