+++ modules/apps/commerce/commerce-product-asset-categories-web/src/main/java/com/liferay/commerce/product/asset/categories/web/internal/portlet/action/UploadTempAssetCategoryAttachmentMVCActionCommand.java
@@ -5,16 +5,44 @@
 
 package com.liferay.commerce.product.asset.categories.web.internal.portlet.action;
 
+import com.liferay.asset.kernel.service.AssetCategoryService;
 import com.liferay.commerce.product.asset.categories.web.internal.constants.CommerceProductAssetCategoriesPortletKeys;
-import com.liferay.commerce.product.asset.categories.web.internal.upload.AssetCategoryAttachmentsUploadResponseHandler;
-import com.liferay.commerce.product.asset.categories.web.internal.upload.TempAssetCategoryAttachmentsUploadFileEntryHandler;
+import com.liferay.commerce.product.configuration.AttachmentsConfiguration;
+import com.liferay.commerce.product.exception.CPAttachmentFileEntryNameException;
+import com.liferay.commerce.product.exception.CPAttachmentFileEntrySizeException;
+import com.liferay.item.selector.ItemSelectorUploadResponseHandler;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.json.JSONUtil;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
+import com.liferay.portal.kernel.repository.model.FileEntry;
+import com.liferay.portal.kernel.servlet.ServletResponseConstants;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.upload.UploadPortletRequest;
+import com.liferay.portal.kernel.util.File;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.TempFileEntryUtil;
+import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.upload.UniqueFileNameProvider;
+import com.liferay.upload.UploadFileEntryHandler;
 import com.liferay.upload.UploadHandler;
+import com.liferay.upload.UploadResponseHandler;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+import java.util.Map;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
+import javax.portlet.PortletRequest;
 
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
@@ -22,6 +50,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
+	configurationPid = "com.liferay.commerce.product.configuration.AttachmentsConfiguration",
 	property = {
 		"javax.portlet.name=" + CommerceProductAssetCategoriesPortletKeys.ASSET_CATEGORIES_ADMIN,
 		"mvc.command.name=/commerce_product_asset_categories/upload_temp_asset_category_attachment"
@@ -30,6 +59,160 @@ import org.osgi.service.component.annotations.Reference;
 )
 public class UploadTempAssetCategoryAttachmentMVCActionCommand
 	extends BaseMVCActionCommand {
+
+	public class AssetCategoryAttachmentsUploadResponseHandler
+		implements UploadResponseHandler {
+
+		@Override
+		public JSONObject onFailure(
+				PortletRequest portletRequest, PortalException portalException)
+			throws PortalException {
+
+			JSONObject jsonObject =
+				_itemSelectorUploadResponseHandler.onFailure(
+					portletRequest, portalException);
+
+			if (portalException instanceof CPAttachmentFileEntryNameException ||
+				portalException instanceof CPAttachmentFileEntrySizeException) {
+
+				String errorMessage = StringPool.BLANK;
+				int errorType = 0;
+
+				if (portalException instanceof
+						CPAttachmentFileEntryNameException) {
+
+					errorType =
+						ServletResponseConstants.SC_FILE_EXTENSION_EXCEPTION;
+
+					errorMessage = StringUtil.merge(
+						_attachmentsConfiguration.imageExtensions());
+				}
+				else if (portalException instanceof
+							CPAttachmentFileEntrySizeException) {
+
+					errorType = ServletResponseConstants.SC_FILE_SIZE_EXCEPTION;
+				}
+
+				jsonObject.put(
+					"error",
+					JSONUtil.put(
+						"errorType", errorType
+					).put(
+						"message", errorMessage
+					));
+			}
+			else {
+				throw portalException;
+			}
+
+			return jsonObject;
+		}
+
+		@Override
+		public JSONObject onSuccess(
+				UploadPortletRequest uploadPortletRequest, FileEntry fileEntry)
+			throws PortalException {
+
+			return _itemSelectorUploadResponseHandler.onSuccess(
+				uploadPortletRequest, fileEntry);
+		}
+
+	}
+
+	public class TempAssetCategoryAttachmentsUploadFileEntryHandler
+		implements UploadFileEntryHandler {
+
+		@Override
+		public FileEntry upload(UploadPortletRequest uploadPortletRequest)
+			throws IOException, PortalException {
+
+			ThemeDisplay themeDisplay =
+				(ThemeDisplay)uploadPortletRequest.getAttribute(
+					WebKeys.THEME_DISPLAY);
+
+			String fileName = uploadPortletRequest.getFileName(_PARAMETER_NAME);
+
+			_validateFile(
+				fileName, uploadPortletRequest.getSize(_PARAMETER_NAME));
+
+			String contentType = uploadPortletRequest.getContentType(
+				_PARAMETER_NAME);
+
+			try (InputStream inputStream = uploadPortletRequest.getFileAsStream(
+					_PARAMETER_NAME)) {
+
+				return _addFileEntry(
+					fileName, contentType, inputStream, themeDisplay);
+			}
+		}
+
+		private FileEntry _addFileEntry(
+				String fileName, String contentType, InputStream inputStream,
+				ThemeDisplay themeDisplay)
+			throws PortalException {
+
+			String uniqueFileName = _uniqueFileNameProvider.provide(
+				fileName, curFileName -> _exists(themeDisplay, curFileName));
+
+			return TempFileEntryUtil.addTempFileEntry(
+				themeDisplay.getScopeGroupId(), themeDisplay.getUserId(),
+				_TEMP_FOLDER_NAME, uniqueFileName, inputStream, contentType);
+		}
+
+		private boolean _exists(ThemeDisplay themeDisplay, String curFileName) {
+			try {
+				FileEntry fileEntry = TempFileEntryUtil.getTempFileEntry(
+					themeDisplay.getScopeGroupId(), themeDisplay.getUserId(),
+					_TEMP_FOLDER_NAME, curFileName);
+
+				if (fileEntry != null) {
+					return true;
+				}
+
+				return false;
+			}
+			catch (PortalException portalException) {
+				if (_log.isDebugEnabled()) {
+					_log.debug(portalException);
+				}
+
+				return false;
+			}
+		}
+
+		private void _validateFile(String fileName, long size)
+			throws PortalException {
+
+			if ((_attachmentsConfiguration.imageMaxSize() > 0) &&
+				(size > _attachmentsConfiguration.imageMaxSize())) {
+
+				throw new CPAttachmentFileEntrySizeException();
+			}
+
+			String extension = _file.getExtension(fileName);
+
+			String[] imageExtensions =
+				_attachmentsConfiguration.imageExtensions();
+
+			for (String imageExtension : imageExtensions) {
+				if (StringPool.STAR.equals(imageExtension) ||
+					imageExtension.equals(StringPool.PERIOD + extension)) {
+
+					return;
+				}
+			}
+
+			throw new CPAttachmentFileEntryNameException(
+				"Invalid image for file name " + fileName);
+		}
+
+	}
+
+	@Activate
+	protected void activate(Map<String, Object> properties) {
+		_attachmentsConfiguration = ConfigurableUtil.createConfigurable(
+			AttachmentsConfiguration.class, properties);
+	}
 
 	@Override
 	protected void doProcessAction(
@@ -43,12 +226,36 @@ public class UploadTempAssetCategoryAttachmentMVCActionCommand
 	}
 
 	@Reference
-	private AssetCategoryAttachmentsUploadResponseHandler
-		_assetCategoryAttachmentsUploadResponseHandler;
+	protected AssetCategoryService assetCategoryService;
+
+	private static final String _PARAMETER_NAME = "imageSelectorFileName";
+
+	private static final String _TEMP_FOLDER_NAME =
+		UploadTempAssetCategoryAttachmentMVCActionCommand.
+			TempAssetCategoryAttachmentsUploadFileEntryHandler.class.getName();
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		UploadTempAssetCategoryAttachmentMVCActionCommand.
+			TempAssetCategoryAttachmentsUploadFileEntryHandler.class);
+
+	private final AssetCategoryAttachmentsUploadResponseHandler
+		_assetCategoryAttachmentsUploadResponseHandler =
+			new AssetCategoryAttachmentsUploadResponseHandler();
+	private volatile AttachmentsConfiguration _attachmentsConfiguration;
 
 	@Reference
-	private TempAssetCategoryAttachmentsUploadFileEntryHandler
-		_tempAssetCategoryAttachmentsUploadFileEntryHandler;
+	private File _file;
+
+	@Reference
+	private ItemSelectorUploadResponseHandler
+		_itemSelectorUploadResponseHandler;
+
+	private final TempAssetCategoryAttachmentsUploadFileEntryHandler
+		_tempAssetCategoryAttachmentsUploadFileEntryHandler =
+			new TempAssetCategoryAttachmentsUploadFileEntryHandler();
+
+	@Reference
+	private UniqueFileNameProvider _uniqueFileNameProvider;
 
 	@Reference
 	private UploadHandler _uploadHandler;
